docker/backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlalchemy
python-dotenv
line-bot-sdk
pymysql
pyngrok
openai==0.28
cryptography
paddlepaddle==3.0.0b2 -f https://www.paddlepaddle.org.cn/packages/stable/cpu
paddleocr>=2.0.1
opencc-python-reimplemented
opencv-python
numpy
python-multipart
passlib[Bcrypt]>=1.7.4
python-jose>=3.3.0
pydantic[email]>=1.10.0
fastapi-mail